+++ sdmx/__init__.py
@@ -4,8 +4,8 @@ import pkg_resources
 
 from sdmx.client import Client, Request, read_url
 from sdmx.reader import read_sdmx
+from sdmx.rest import Resource
 from sdmx.source import add_source, list_sources
-from sdmx.util import Resource
 from sdmx.writer import to_pandas, to_xml
 
 __all__ = [
@@ -24,7 +24,7 @@ __all__ = [
 
 try:
     __version__ = pkg_resources.get_distribution("sdmx1").version
-except Exception:
+except Exception:  # pragma: no cover
     # Local copy or not installed with setuptools
     __version__ = "999"