+++ sdmx/source/__init__.py
@@ -1,13 +1,15 @@
 import json
 from enum import Enum
 from importlib import import_module
-from io import TextIOWrapper
-from typing import Any, Dict, Union
+from io import IOBase, TextIOWrapper
+from typing import Any, Dict, Optional, Tuple, Union
 
 from pkg_resources import resource_stream
+from requests import Response
 
 from sdmx.model import DataStructureDefinition
-from sdmx.util import BaseModel, Resource, validator
+from sdmx.rest import Resource
+from sdmx.util import BaseModel, validator
 
 sources: Dict[str, "Source"] = {}
 
@@ -17,9 +19,9 @@ DataContentType = Enum("DataContentType", "XML JSON")
 class Source(BaseModel):
     """SDMX-IM RESTDatasource.
 
-    This class describes the location and features supported by an SDMX data
-    source. Subclasses may override the hooks in order to handle specific
-    features of different REST web services:
+    This class describes the location and features supported by an SDMX data source.
+    Subclasses may override the hooks in order to handle specific features of different
+    REST web services:
 
     .. autosummary::
        handle_response
@@ -39,25 +41,54 @@ class Source(BaseModel):
 
     headers: Dict[str, Any] = {}
 
-    #: :class:`.DataContentType` indicating the type of data returned by the
-    #: source.
+    #: :class:`.DataContentType` indicating the type of data returned by the source.
     data_content_type: DataContentType = DataContentType.XML
 
-    #: Mapping from :class:`~sdmx.Resource` to :class:`bool` indicating support
-    #: for SDMX REST API features. Two additional keys are valid:
+    #: Mapping from :class:`.Resource` values to :class:`bool` indicating support for
+    #: SDMX-REST endpoints and features. Most of these values indicate endpoints that
+    #: are described in the standards but are not implemented by any source currently in
+    #: :file:`sources.json`; these all return 404.
+    #:
+    #: Two additional keys are valid:
     #:
     #: - ``'preview'=True`` if the source supports ``?detail=serieskeysonly``.
     #:   See :meth:`.preview_data`.
     #: - ``'structure-specific data'=True`` if the source can return structure-
     #:   specific data messages.
-    supports: Dict[Union[str, Resource], bool] = {Resource.data: True}
+    supports: Dict[Union[str, Resource], bool] = {
+        Resource.data: True,
+        Resource.actualconstraint: False,
+        Resource.allowedconstraint: False,
+        Resource.attachementconstraint: False,
+        Resource.customtypescheme: False,
+        Resource.dataconsumerscheme: False,
+        Resource.dataproviderscheme: False,
+        Resource.hierarchicalcodelist: False,
+        Resource.metadata: False,
+        Resource.metadataflow: False,
+        Resource.metadatastructure: False,
+        Resource.namepersonalisationscheme: False,
+        Resource.organisationunitscheme: False,
+        Resource.process: False,
+        Resource.reportingtaxonomy: False,
+        Resource.rulesetscheme: False,
+        Resource.schema: False,
+        Resource.transformationscheme: False,
+        Resource.userdefinedoperatorscheme: False,
+        Resource.vtlmappingscheme: False,
+    }
 
     @classmethod
     def from_dict(cls, info):
         return cls(**info)
 
     def __init__(self, **kwargs):
+        # Merge supports values with defaults
+        supports = kwargs.pop("supports", dict())
+
         super().__init__(**kwargs)
+
+        self.supports.update(supports)
 
         # Set default supported features
         for feature in list(Resource) + ["preview", "structure-specific data"]:
@@ -66,7 +97,9 @@ class Source(BaseModel):
             )
 
     # Hooks
-    def handle_response(self, response, content):
+    def handle_response(
+        self, response: Response, content: IOBase
+    ) -> Tuple[Response, IOBase]:
         """Handle response content of unknown type.
 
         This hook is called by :meth:`.Client.get` *only* when the `content` cannot be
@@ -134,7 +167,9 @@ class _NoSource(Source):
 NoSource = _NoSource()
 
 
-def add_source(info, id=None, override=False, **kwargs):
+def add_source(
+    info: Union[Dict, str], id: Optional[str] = None, override: bool = False, **kwargs
+) -> None:
     """Add a new data source.
 
     The *info* expected is in JSON format:
@@ -146,35 +181,32 @@ def add_source(info, id=None, override=False, **kwargs):
           "documentation": "http://data.un.org/Host.aspx?Content=API",
           "url": "http://ec.europa.eu/eurostat/SDMX/diss-web/rest",
           "name": "Eurostat",
-          "supported": {"codelist": false, "preview": true}
+          "supports": {"codelist": false, "preview": true}
         }
 
-    …with unspecified values using the defaults; see
-    :class:`Source`.
+    …with unspecified values using the defaults; see :class:`Source`.
 
     Parameters
     ----------
     info : dict-like
         String containing JSON information about a data source.
     id : str
-        Identifier for the new datasource. If :obj:`None` (default), then
-        `info['id']` is used.
+        Identifier for the new datasource. If :obj:`None` (default), then `info['id']`
+        is used.
     override : bool
-        If :obj:`True`, replace any existing data source with *id*.
-        Otherwise, raise :class:`ValueError`.
+        If :obj:`True`, replace any existing data source with *id*. Otherwise, raise
+        :class:`ValueError`.
     **kwargs
         Optional callbacks for *handle_response* and *finish_message* hooks.
 
     """
-    if isinstance(info, str):
-        info = json.loads(info)
+    _info = json.loads(info) if isinstance(info, str) else info
+    id = id or _info["id"]
 
-    id = info["id"] if id is None else id
-
-    info.update(kwargs)
+    _info.update(kwargs)
 
     if not override and id in sources:
-        raise ValueError("Data source '%s' already defined; use override=True", id)
+        raise ValueError(f"Data source {repr(id)} already defined; use override=True")
 
     # Maybe import a subclass that defines a hook
     SourceClass = Source
@@ -183,9 +215,9 @@ def add_source(info, id=None, override=False, **kwargs):
     except ImportError:
         pass
     else:
-        SourceClass = mod.Source
+        SourceClass = getattr(mod, "Source")
 
-    sources[id] = SourceClass.from_dict(info)
+    sources[id] = SourceClass.from_dict(_info)
 
 
 def list_sources():
@@ -197,7 +229,7 @@ def list_sources():
 
 
 def load_package_sources():
-    """Discover all sources listed in ``sources.json``."""
+    """Discover all sources listed in :file:`sources.json`."""
     with resource_stream("sdmx", "sources.json") as f:
         # TextIOWrapper is for Python 3.5 compatibility
         for info in json.load(TextIOWrapper(f)):