+++ sdmx/urn.py
@@ -18,14 +18,14 @@ _BASE = (
 )
 
 
-def make(obj, maintainable_parent=None):
+def make(obj, maintainable_parent=None, strict=False):
     """Create an SDMX URN for `obj`.
 
     If `obj` is not :class:`.MaintainableArtefact`, then `maintainable_parent`
     must be supplied in order to construct the URN.
     """
-    if maintainable_parent:
-        ma = maintainable_parent
+    if not isinstance(obj, MaintainableArtefact):
+        ma = maintainable_parent or obj.get_scheme()
         extra_id = f".{obj.id}"
     else:
         ma = obj
@@ -36,7 +36,9 @@ def make(obj, maintainable_parent=None):
             f"Neither {repr(obj)} nor {repr(maintainable_parent)} are maintainable"
         )
     elif ma.maintainer is None:
-        raise ValueError(f"Cannot construct URL for {repr(ma)} without maintainer")
+        raise ValueError(f"Cannot construct URN for {repr(ma)} without maintainer")
+    elif strict and ma.version is None:
+        raise ValueError(f"Cannot construct URN for {repr(ma)} without version")
 
     return _BASE.format(
         package=PACKAGE[obj.__class__], obj=obj, ma=ma, extra_id=extra_id